==> 00.DataDownloader.ipynb <==
import wrds
import pandas as pd
conn = wrds.Connection()

query = """
    SELECT gvkey, datadate, fyear, at, sale, ib, cogs, xsga, txt, dp, oiadp, spi
    FROM comp.funda
    WHERE indfmt = 'INDL'
    AND datafmt = 'STD'
    AND popsrc = 'D'
    AND consol = 'C'
    AND fyear >= 1950
    AND fyear <= 2023
"""

data = conn.raw_sql(query)

data.to_pickle('./data/raw_data.pkl', compression='zip')
conn.close()
