==> 02.1.Table1.ipynb <==
import warnings
import pandas as pd
import numpy as np
import statsmodels.api as sm
from scipy.stats import ttest_ind
warnings.filterwarnings('ignore')

start_y = 1967
end_y = 2005
split_y = 1985

data = pd.read_csv('./data/master_data.csv')
data = data.loc[(data['fyear'] >= start_y) & (data['fyear'] <= end_y)]
data = data.dropna(subset=['exp_lag', 'exp', 'exp_lead', 'earn'])

def t_test(before, after, coef):
    t_stat, p_value = ttest_ind(before[coef], after[coef])
    return t_stat, p_value

# Panel A: Cross sectional regressions
results = pd.DataFrame(columns=['exp_lag', 'exp_cur', 'exp_lead'])


for year, group in data.groupby('fyear'):
    X = group[['exp_lag', 'exp', 'exp_lead']]
    X = sm.add_constant(X)
    y = group['rev']

    model = sm.OLS(y, X).fit()
    coef = model.params

    result = pd.DataFrame(
        {
            'exp_lag': coef['exp_lag'],
            'exp_cur': coef['exp'],
            'exp_lead': coef['exp_lead']
        },
        index=[year]
    )
    results = pd.concat([results, result])


results = results.reset_index(names='year')

panel_a = pd.DataFrame(columns=['exp_lag', 'exp_cur', 'exp_lead'])
panel_a = pd.concat([panel_a, results.loc[results['year'] <= split_y, ['exp_lag', 'exp_cur', 'exp_lead']].mean().to_frame().T], ignore_index=True)
panel_a = pd.concat([panel_a, results.loc[results['year'] >= split_y + 1, ['exp_lag', 'exp_cur', 'exp_lead']].mean().to_frame().T], ignore_index=True)
panel_a.loc['difference'] = panel_a.iloc[1] - panel_a.iloc[0]
panel_a.loc['t-statistic'] = np.nan
panel_a.loc['p-value'] = np.nan


for coef in ['exp_lag', 'exp_cur', 'exp_lead']:
    t, p = t_test(results.loc[results['year'] <= split_y], results.loc[results['year'] >= split_y + 1], coef)
    panel_a.loc['t-statistic', coef] = t
    panel_a.loc['p-value', coef] = p

panel_a

# Panel B: Earn Volat and Persistence
persistence = pd.DataFrame(columns=['persistence'])


for year, group in data.groupby('fyear'):
    X = group[['earn_lag']]
    X = sm.add_constant(X)
    y = group['earn']

    model = sm.OLS(y, X).fit()
    coef = model.params

    result = pd.DataFrame(
        {
            'persistence': coef['earn_lag']
        },
        index=[year]
    )
    persistence = pd.concat([persistence, result])


persistence = persistence.reset_index(names='year')
volatility = data.groupby('fyear')['earn_5yr_volat'].mean().reset_index()

panel_b = pd.DataFrame(columns=['volatility', 'persistence'])
panel_b.loc[0, 'volatility'] = volatility.loc[volatility['fyear'] <= split_y, 'earn_5yr_volat'].mean()
panel_b.loc[1, 'volatility'] = volatility.loc[volatility['fyear'] >= split_y + 1, 'earn_5yr_volat'].mean()
panel_b.loc[0, 'persistence'] = persistence.loc[persistence['year'] <= split_y, 'persistence'].mean()
panel_b.loc[1, 'persistence'] = persistence.loc[persistence['year'] >= split_y + 1, 'persistence'].mean()
panel_b.loc['difference'] = panel_b.iloc[1] - panel_b.iloc[0]
panel_b.loc['t-statistic'] = np.nan
panel_b.loc['p-value'] = np.nan
panel_b.loc[['t-statistic', 'p-value'], 'volatility'] = t_test(volatility.loc[volatility['fyear'] <= split_y], volatility.loc[volatility['fyear'] >= split_y + 1], 'earn_5yr_volat')
panel_b.loc[['t-statistic', 'p-value'], 'persistence'] = t_test(persistence.loc[persistence['year'] <= split_y], persistence.loc[persistence['year'] >= split_y + 1], 'persistence')

panel_b
